Improve board drawing.
There were problems when
the board was drawn at a
size not divisible by 8.
What happens, sometimes.

This fix is not perfect,
as it makes resizing the
board a bit jumpy. It is
impossible to do better,
as long as there is such
a side column; that will
be fixed for next cycle.