game-board-view: Don't assume the current event is a button press
When using a touchscreen, the current event will be of type TOUCH_BEGIN
rather than BUTTON_PRESS, and this populates different fields in the
GdkEvent union.

The GtkGestureMultiPress has already processed the click/touch/etc.
event and derived x and y coordinates from the raw data, so all we
need to do is to round them to an integer.

Resolves: https://gitlab.gnome.org/GNOME/four-in-a-row/-/issues/27
Bug-Debian: https://bugs.debian.org/1009749
Signed-off-by: Simon McVittie <smcv@debian.org>